Changed Like Dislike buttons
Changed the location to already made templates for easier logging
